requirements.txt
Django>=5.0
djangorestframework>=3.15
celery>=5.3
redis>=5.0
opencv-python-headless>=4.9
numpy>=1.26
Pillow>=10.2
